gmail_otp/requirements.txt
fastapi
uvicorn
strawberry-graphql
python-dotenv


